refactor(SF-133): extract route-telemetry helpers from routes_shared.py
routes_shared.py is the shared FastAPI router factory for every
*_backend_api plugin. It mixed FastAPI wiring with five OTEL
recording helpers; the helpers are pure side-effect utilities with
no FastAPI / closure dependency.

Extract to:
- llm_base/_route_telemetry.py (98 lines): record_ignored_fields,
  record_span_success, record_url4_span, plus the _otel_attr_value
  coercion helper.

routes_shared.py: 456 → 380 lines (-17%). Public API
(BackendApiConfig + build_backend_api_router) unchanged. 722 plugin
unit tests pass.

https://app.asana.com/1/1185126988600652/project/1213628819033917/task/1214358855646473
